Navigation: Extract NavigationLinkUI and NavigationListViewHeader into separate files
Refactor menu-inspector-controls.js to extract two components into their own files:
- NavigationLinkUI (formerly AdditionalBlockContent) → navigation-link-ui.js
- NavigationListViewHeader → navigation-list-view-header.js

This is a pure refactor with no behavior changes, intended to simplify the review of #75694.
